requirements.txt
pyyaml
neo4j
openai>=1.37.0
tiktoken
fastapi
uvicorn[standard]
celery[redis]
redis
python-multipart
